+++ formula-test/src/main/java/com/instacart/formula/test/TestFormulaObserver.kt
@@ -7,7 +7,6 @@ import com.instacart.formula.plugin.Inspector
 import com.instacart.formula.toFlow
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
